requirements.txt
annotated-types==0.7.0
anyio==4.8.0
attrs==25.3.0
beautifulsoup4==4.13.3
bs4==0.0.2
certifi==2025.1.31
charset-normalizer==3.4.1
coverage==7.9.2
distro==1.9.0
dotenv==0.9.9
h11==0.14.0
html5lib==1.1
httpcore==1.0.7
httpx==0.28.1
hypothesis==6.135.27
idna==3.10
iniconfig==2.1.0
jiter==0.8.2
lxml==5.3.1
openai==1.65.3
packaging==25.0
pluggy==1.6.0
pydantic==2.10.6
pydantic-core==2.27.2
pygments==2.19.2
pytest==8.4.1
pytest-cov==6.2.1
pytest-mock==3.14.1
python-docx==1.1.2
python-dotenv==1.0.1
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.6
tqdm==4.67.1
typing-extensions==4.12.2
urllib3==2.3.0
webencodings==0.5.1
